+++ src/pudl/transform/eia860.py
@@ -357,6 +357,60 @@ def _core_eia860__generators(
 
 
 @asset
+def _core_eia860__generators_solar(
+    raw_eia860__generator_solar_existing: pd.DataFrame,
+    raw_eia860__generator_solar_retired: pd.DataFrame,
+) -> pd.DataFrame:
+    """Transform the solar-specific generators table.
+
+    Many of the same transforms to the core generators table are applied here.
+    Most of the unique solar columns are booleans.
+
+    Notes for possible future cleaning:
+
+    * Both the ``tilt_angle`` and ``azimuth_angle`` columns have a small number of
+      negative values (both under 40 records). This seems off, but not impossible?
+    * A lot of the boolean columns in this table are mostly null. It is probably
+      that a lot of the nulls should coorespond to False's, but there is no sure way
+      to know, so nulls seem more appropriate.
+
+    """
+    solar_existing = raw_eia860__generator_solar_existing
+    solar_retired = raw_eia860__generator_solar_retired
+    # every boolean column in the raw solar tables has a uses prefix
+    boolean_columns_to_fix = list(solar_existing.filter(regex=r"^uses_"))
+    if mismatched_bool_cols := set(boolean_columns_to_fix).difference(
+        set(solar_retired.filter(regex=r"^uses_"))
+    ):
+        raise AssertionError(
+            "We expect that the raw existing and retired assets to have the exact same "
+            f"boolean columns with prefix of uses_ but we found {mismatched_bool_cols=}"
+        )
+    solar_df = (
+        pd.concat([solar_existing, solar_retired], sort=True)
+        .pipe(pudl.helpers.fix_eia_na)
+        .pipe(fix_boolean_columns, boolean_columns_to_fix)
+        .pipe(pudl.helpers.month_year_to_date)
+        .pipe(pudl.helpers.convert_to_date)
+        .pipe(
+            pudl.metadata.classes.Package.from_resource_ids()
+            .get_resource("core_eia860__scd_generators")
+            .encode
+        )
+    )
+
+    solar_df["operational_status"] = solar_df.operational_status_code.str.upper().map(
+        pudl.helpers.label_map(
+            CODE_METADATA["core_eia__codes_operational_status"]["df"],
+            from_col="code",
+            to_col="operational_status",
+            null_value=pd.NA,
+        )
+    )
+    return solar_df
+
+
+@asset
 def _core_eia860__generators_energy_storage(
     raw_eia860__generator_energy_storage_existing: pd.DataFrame,
     raw_eia860__generator_energy_storage_retired: pd.DataFrame,
@@ -381,11 +435,6 @@ def _core_eia860__generators_energy_storage(
         .pipe(pudl.helpers.month_year_to_date)
         .pipe(pudl.helpers.convert_to_date)
         .pipe(fix_boolean_columns, boolean_columns_to_fix=boolean_columns_to_fix)
-        .pipe(
-            pudl.metadata.classes.Package.from_resource_ids()
-            .get_resource("core_eia860__scd_generators")
-            .encode
-        )
         .pipe(
             pudl.metadata.classes.Package.from_resource_ids()
             .get_resource("core_eia860__yearly_generators_energy_storage")
@@ -451,7 +500,7 @@ def _core_eia860__generators_wind(
         )
         .pipe(
             pudl.metadata.classes.Package.from_resource_ids()
-            .get_resource("core_eia860__yearly_generators_wind")
+            .get_resource("core_eia860__scd_generators_wind")
             .encode
         )
     )
@@ -673,8 +722,9 @@ def _core_eia860__utilities(raw_eia860__utility: pd.DataFrame) -> pd.DataFrame:
 
 @asset
 def _core_eia860__boilers(
-    raw_eia860__emission_control_strategies, raw_eia860__boiler_info
-):
+    raw_eia860__emission_control_strategies: pd.DataFrame,
+    raw_eia860__boiler_info: pd.DataFrame,
+) -> pd.DataFrame:
     """Pull and transform the boilers table.
 
     Transformations include:
@@ -687,13 +737,13 @@ def _core_eia860__boilers(
       reporting.
 
     Args:
-        raw_eia860__emission_control_strategies (pandas.DataFrame):
-            DataFrame extracted from EIA forms earlier in the ETL process.
-        raw_eia860__boiler_info (pandas.DataFrame):
-            DataFrame extracted from EIA forms earlier in the ETL process.
+        raw_eia860__emission_control_strategies: DataFrame extracted from EIA forms
+            earlier in the ETL process.
+        raw_eia860__boiler_info: DataFrame extracted from EIA forms earlier in the ETL
+            process.
 
     Returns:
-        pandas.DataFrame: the transformed boilers table
+        The transformed boilers table.
     """
     # Populating the 'core_eia860__scd_boilers' table
     b_df = raw_eia860__boiler_info
@@ -885,7 +935,9 @@ def _core_eia860__emissions_control_equipment(
     emce_df = pudl.helpers.fix_eia_na(raw_eia860__emissions_control_equipment)
 
     # Spot fix bad months
-    emce_df["operating_month"] = emce_df["operating_month"].replace({"88": "8"})
+    emce_df["emission_control_operating_month"] = emce_df[
+        "emission_control_operating_month"
+    ].replace({"88": "8"})
     # Fill in values with a year not a month based on later years that do have a month
     # I thought about doing some sort of backfill here, but decided not to because
     # emission_control_id_pudl is not guaranteed to be consistent over time
@@ -902,7 +954,7 @@ def _core_eia860__emissions_control_equipment(
     bad_month_3 = (
         (emce_df["report_year"] == 2013)
         & (emce_df["plant_id_eia"] == 3131)
-        & (emce_df["operating_year"] == "2005")
+        & (emce_df["emission_control_operating_year"] == "2005")
     )
     bad_month_4 = (emce_df["report_year"] == 2013) & (emce_df["plant_id_eia"] == 10405)
     bad_month_5 = (
@@ -913,12 +965,12 @@ def _core_eia860__emissions_control_equipment(
     bad_month_6 = (
         (emce_df["report_year"] == 2013)
         & (emce_df["plant_id_eia"] == 4054)
-        & (emce_df["operating_year"] == "2011")
+        & (emce_df["emission_control_operating_year"] == "2011")
     )
     bad_month_7 = (
         (emce_df["report_year"] == 2013)
         & (emce_df["plant_id_eia"] == 50544)
-        & (emce_df["operating_year"] == "1990")
+        & (emce_df["emission_control_operating_year"] == "1990")
     )
     bad_month_8 = (
         (emce_df["report_year"] == 2013)
@@ -930,33 +982,33 @@ def _core_eia860__emissions_control_equipment(
     # (in which case the assertions will fail)
     if 2013 in emce_df.report_year.unique():
         assert len(emce_df[bad_month_1]) == 1
-        emce_df.loc[bad_month_1, "operating_month"] = 6
+        emce_df.loc[bad_month_1, "emission_control_operating_month"] = 6
         assert len(emce_df[bad_month_2]) == 4
-        emce_df.loc[bad_month_2, "operating_month"] = 6
+        emce_df.loc[bad_month_2, "emission_control_operating_month"] = 6
         assert len(emce_df[bad_month_3]) == 4
-        emce_df.loc[bad_month_3, "operating_month"] = 1
+        emce_df.loc[bad_month_3, "emission_control_operating_month"] = 1
         assert len(emce_df[bad_month_4]) == 1
-        emce_df.loc[bad_month_4, "operating_month"] = 12
+        emce_df.loc[bad_month_4, "emission_control_operating_month"] = 12
         assert len(emce_df[bad_month_5]) == 2
-        emce_df.loc[bad_month_5, "operating_month"] = 6
+        emce_df.loc[bad_month_5, "emission_control_operating_month"] = 6
         assert len(emce_df[bad_month_6]) == 2
-        emce_df.loc[bad_month_6, "operating_month"] = 10
+        emce_df.loc[bad_month_6, "emission_control_operating_month"] = 10
         assert len(emce_df[bad_month_7]) == 1
-        emce_df.loc[bad_month_7, "operating_month"] = 6
+        emce_df.loc[bad_month_7, "emission_control_operating_month"] = 6
         assert len(emce_df[bad_month_8]) == 9
-        emce_df.loc[bad_month_8, "operating_month"] = 12
+        emce_df.loc[bad_month_8, "emission_control_operating_month"] = 12
 
     # Convert month-year columns to a single date column
     emce_df = pudl.helpers.convert_to_date(
         df=emce_df,
         date_col="emission_control_operating_date",
-        year_col="operating_year",
-        month_col="operating_month",
+        year_col="emission_control_operating_year",
+        month_col="emission_control_operating_month",
     ).pipe(
         pudl.helpers.convert_to_date,
         date_col="emission_control_retirement_date",
-        year_col="retirement_year",
-        month_col="retirement_month",
+        year_col="emission_control_retirement_year",
+        month_col="emission_control_retirement_month",
     )
 
     # Convert acid gas control column to boolean
@@ -1024,8 +1076,7 @@ def _core_eia860__boiler_emissions_control_equipment_assn(
             association table.
 
     Returns:
-        pd.DataFrame: A combination of all the emission control equipment association
-            tables.
+        A combination of all the emission control equipment association tables.
     """
     raw_tables = [
         raw_eia860__boiler_so2,
@@ -1090,8 +1141,7 @@ def _core_eia860__boiler_cooling(
         raw_eia860__boiler_cooling: Raw EIA 860 boiler to cooler ID association table.
 
     Returns:
-        pd.DataFrame: A cleaned and normalized version of the EIA boiler to cooler ID
-            table.
+        A cleaned and normalized version of the EIA boiler to cooler ID table.
     """
     # Replace empty strings, whitespace, and '.' fields with real NA values
     bc_assn = pudl.helpers.fix_eia_na(raw_eia860__boiler_cooling)
@@ -1116,8 +1166,7 @@ def _core_eia860__boiler_stack_flue(
             table.
 
     Returns:
-        pd.DataFrame: A cleaned and normalized version of the EIA boiler to stack flue
-            ID table.
+        A cleaned and normalized version of the EIA boiler to stack flue ID table.
     """
     # Replace empty strings, whitespace, and '.' fields with real NA values
     bsf_assn = pudl.helpers.fix_eia_na(raw_eia860__boiler_stack_flue)
@@ -1188,10 +1237,8 @@ def _core_eia860__cooling_equipment(
     ] = 1971
     ce_df.loc[ce_df["pond_operating_year"] == 0, "pond_operating_year"] = np.nan
     ce_df.loc[ce_df["tower_operating_year"] == 974, "tower_operating_year"] = 1974
-    ce_df = (
-        ce_df.pipe(pudl.helpers.month_year_to_date)
-        .pipe(pudl.helpers.convert_to_date)
-        .rename(columns={"operating_date": "cooling_system_operating_date"})
+    ce_df = ce_df.pipe(pudl.helpers.month_year_to_date).pipe(
+        pudl.helpers.convert_to_date
     )
 
     # There's one row which has an NA cooling_id_eia, which we mark as "PLANT"
@@ -1304,7 +1351,6 @@ def _core_eia860__fgd_equipment(
     fgd_df = fgd_df.pipe(pudl.helpers.month_year_to_date).pipe(
         pudl.helpers.convert_to_date
     )
-    fgd_df = fgd_df.rename(columns={"operating_date": "fgd_operating_date"})
 
     # Handle mixed boolean types in control flag column
     for col in [
@@ -1388,7 +1434,7 @@ def fgd_equipment_null_check(fgd):  # pragma: no cover
     fast_run_null_cols = {
         "county",
         "fgd_operational_status_code",
-        "operating_date",
+        "fgd_operating_date",
         "fgd_manufacturer_code",
         "plant_summer_capacity_mw",
         "water_source",